Expose DataStorePredicateObject type
