+++ packages/react/src/primitives/types/index.ts
@@ -6,6 +6,7 @@ export * from './buttonGroup';
 export * from './card';
 export * from './checkboxField';
 export * from './collection';
+export * from './datastore';
 export * from './divider';
 export * from './field';
 export * from './fieldGroup';
